Move `Agent`->`RecordBatch` conversion to `stateful`
