+++ packages/engine/lib/stateful/src/agent/arrow/mod.rs
@@ -1,0 +1,5 @@
+pub mod array;
+
+// TODO: this should be deleted, i.e. if this value is required use
+//      something like `get_hidden_column_name(PREVIOUS_INDEX_FIELD_NAME)`
+pub const PREVIOUS_INDEX_FIELD_KEY: &str = "_HIDDEN_0_previous_index";
