+++ packages/engine/src/datastore/arrow/batch_conversion.rs
@@ -7,257 +7,28 @@
 use std::{collections::HashSet, sync::Arc};
 
 use arrow::{
-    array::{self, Array, ArrayData, ArrayDataBuilder, ArrayRef, FixedSizeListArray, StringArray},
-    buffer::MutableBuffer,
-    datatypes::{self, DataType, Field, Schema},
+    array::{self, Array, FixedSizeListArray, StringArray},
+    datatypes::Field,
     record_batch::RecordBatch,
-    util::bit_util,
 };
-use memory::arrow::{
-    col_to_json_vals, json_utf8_json_vals, json_vals_to_any_type_col, json_vals_to_bool,
-    json_vals_to_col, json_vals_to_primitive, json_vals_to_utf8, new_zero_bits,
-};
-use serde_json::value::Value;
+use memory::arrow::{col_to_json_vals, json_utf8_json_vals};
 use stateful::{
-    agent::{Agent, AgentName, AgentSchema, AgentStateField, BUILTIN_FIELDS},
-    field::{FieldScope, FieldTypeVariant, RootFieldKey},
-    message::{
-        arrow::{array::MessageArray, column::MessageColumn},
-        MessageSchema,
+    agent::{
+        arrow::PREVIOUS_INDEX_FIELD_KEY, Agent, AgentName, AgentSchema, AgentStateField,
+        BUILTIN_FIELDS,
     },
+    field::{FieldScope, FieldTypeVariant},
+    message::{arrow::column::MessageColumn, MessageSchema},
+    Error, Result,
 };
 
-use crate::{
-    datastore::{
-        arrow::message,
-        batch::{AgentBatch, MessageBatch},
-        error::{Error, Result},
-        schema::IsRequired,
-        UUID_V4_LEN,
-    },
-    simulation::package::creator::PREVIOUS_INDEX_FIELD_KEY,
+use crate::datastore::{
+    batch::{AgentBatch, MessageBatch},
+    schema::IsRequired,
+    UUID_V4_LEN,
 };
 
 // This file is here mostly to convert between RecordBatch and Vec<Agent>.
-
-/// Conversion into Arrow `RecordBatch`
-pub trait IntoRecordBatch {
-    fn into_message_batch(&self, schema: &Arc<Schema>) -> Result<RecordBatch>;
-    fn into_empty_message_batch(&self, schema: &Arc<Schema>) -> Result<RecordBatch>;
-    /// TODO: DOC describe, explain self is initialization data
-    fn into_agent_batch(&self, schema: &Arc<AgentSchema>) -> Result<RecordBatch>;
-}
-
-fn builder_add_id(builder: &mut array::FixedSizeBinaryBuilder, id: &str) -> Result<()> {
-    if id.is_empty() {
-        // Generates UUID if it does not exist
-        let uuid = uuid::Uuid::new_v4();
-        let bytes = uuid.as_bytes();
-        builder.append_value(bytes)?;
-    } else if let Ok(uuid) = uuid::Uuid::parse_str(id) {
-        builder.append_value(uuid.as_bytes())?;
-    } else {
-        return Err(Error::InvalidAgentId(id.into()));
-    }
-
-    Ok(())
-}
-
-pub fn get_agent_id_array(values: Vec<&str>) -> Result<array::FixedSizeBinaryArray> {
-    let mut builder =
-        array::FixedSizeBinaryBuilder::new(values.len() * UUID_V4_LEN, UUID_V4_LEN as i32);
-    for value in values {
-        builder_add_id(&mut builder, value)?;
-    }
-    Ok(builder.finish())
-}
-
-// `get_agent_id_array` is needed for public interface, but
-// this function avoids copying ids to separate `Vec`.
-fn agents_to_id_col(agents: &[&Agent]) -> Result<ArrayRef> {
-    let mut builder =
-        array::FixedSizeBinaryBuilder::new(agents.len() * UUID_V4_LEN, UUID_V4_LEN as i32);
-    for agent in agents {
-        builder_add_id(&mut builder, &agent.agent_id)?;
-    }
-    Ok(Arc::new(builder.finish()))
-}
-
-macro_rules! agents_to_vec_col_gen {
-    ($field_name:ident, $function_name:ident) => {
-        fn $function_name(agents: &[&Agent]) -> Result<FixedSizeListArray> {
-            let mut flat: Vec<f64> = Vec::with_capacity(agents.len() * 3);
-            let mut null_bits = new_zero_bits(agents.len());
-            let mut_null_bits = null_bits.as_slice_mut();
-            let mut null_count = 0;
-            for (i_agent, agent) in agents.iter().enumerate() {
-                if let Some(dir) = agent.$field_name {
-                    flat.push(dir.0);
-                    flat.push(dir.1);
-                    flat.push(dir.2);
-                    bit_util::set_bit(mut_null_bits, i_agent);
-                } else {
-                    // Null -- put arbitrary data
-                    flat.push(0.0);
-                    flat.push(0.0);
-                    flat.push(0.0);
-                    null_count += 1;
-                }
-            }
-            let child_array: array::Float64Array = flat.into();
-
-            let dt =
-                DataType::FixedSizeList(Box::new(Field::new("item", DataType::Float64, true)), 3);
-
-            Ok(ArrayData::builder(dt)
-                .len(agents.len())
-                .null_count(null_count)
-                .null_bit_buffer(null_bits.into())
-                .add_child_data(child_array.data().clone())
-                .build()?
-                .into())
-        }
-    };
-}
-
-agents_to_vec_col_gen!(direction, agents_to_direction_col);
-agents_to_vec_col_gen!(position, agents_to_position_col);
-agents_to_vec_col_gen!(scale, agents_to_scale_col);
-agents_to_vec_col_gen!(rgb, agents_to_rgb_col);
-
-fn previous_index_to_empty_col(num_agents: usize, dt: DataType) -> Result<ArrayRef> {
-    if let DataType::FixedSizeList(inner_field, inner_len) = dt.clone() {
-        debug_assert!(matches!(inner_field.data_type(), DataType::UInt32));
-        let data_byte_size = inner_len as usize * num_agents * std::mem::size_of::<u32>();
-        let mut buffer = MutableBuffer::new(data_byte_size);
-        buffer.resize(data_byte_size, 0);
-
-        let data = ArrayDataBuilder::new(dt)
-            .len(num_agents)
-            .add_child_data(
-                ArrayData::builder(inner_field.data_type().clone())
-                    .len(num_agents * inner_len as usize)
-                    .add_buffer(buffer.into())
-                    .build()?,
-            )
-            .build()?;
-
-        Ok(Arc::new(arrow::array::FixedSizeListArray::from(data)))
-    } else {
-        Err(Error::from(format!(
-            "previous_index_to_empty_col was called on the wrong datatype: {:?}",
-            dt
-        )))
-    }
-}
-
-impl IntoRecordBatch for &[Agent] {
-    fn into_message_batch(&self, schema: &Arc<Schema>) -> Result<RecordBatch> {
-        self.iter()
-            .collect::<Vec<_>>()
-            .as_slice()
-            .into_message_batch(schema)
-    }
-
-    fn into_empty_message_batch(&self, schema: &Arc<Schema>) -> Result<RecordBatch> {
-        self.iter()
-            .collect::<Vec<_>>()
-            .as_slice()
-            .into_empty_message_batch(schema)
-    }
-
-    fn into_agent_batch(&self, schema: &Arc<AgentSchema>) -> Result<RecordBatch> {
-        self.iter()
-            .collect::<Vec<_>>()
-            .as_slice()
-            .into_agent_batch(schema)
-    }
-}
-
-impl IntoRecordBatch for &[&Agent] {
-    fn into_message_batch(&self, schema: &Arc<Schema>) -> Result<RecordBatch> {
-        let ids = self
-            .iter()
-            .map(|agent| agent.agent_id.as_ref())
-            .collect::<Vec<&str>>();
-        let messages: Vec<Value> = self
-            .iter()
-            .map(|agent| agent.get_as_json("messages"))
-            .collect::<stateful::Result<_>>()?;
-
-        message::batch_from_json(schema, ids, Some(messages))
-    }
-
-    fn into_empty_message_batch(&self, schema: &Arc<Schema>) -> Result<RecordBatch> {
-        let ids = self
-            .iter()
-            .map(|agent| agent.agent_id.as_ref())
-            .collect::<Vec<&str>>();
-        message::batch_from_json(schema, ids, None)
-    }
-
-    fn into_agent_batch(&self, schema: &Arc<AgentSchema>) -> Result<RecordBatch> {
-        let mut cols = Vec::with_capacity(schema.arrow.fields().len());
-
-        for field in schema.arrow.fields() {
-            // If `name` isn't cloned, Rust wants schema to have longer lifetime.
-            let name = field.name().clone();
-
-            let vals: Vec<Value> = self
-                .iter()
-                .map(|agent: &&Agent| agent.get_as_json(name.as_str()))
-                .collect::<stateful::Result<_>>()?;
-
-            // If use `match` instead of `if`, Rust infers that
-            // `name` must have static lifetime, like `match` arms.
-            // TODO: built-ins should take nullability from the schema
-            let col = if name.eq(AgentStateField::AgentId.name()) {
-                agents_to_id_col(*self)?
-            } else if name == AgentStateField::AgentName.name() {
-                Arc::new(json_vals_to_utf8(vals, true)?)
-            } else if name == AgentStateField::Messages.name() {
-                Arc::new(MessageArray::from_json(vals)?)
-            } else if name == AgentStateField::Position.name() {
-                Arc::new(agents_to_position_col(*self)?)
-            } else if name == AgentStateField::Direction.name()
-                || name == AgentStateField::Velocity.name()
-            {
-                Arc::new(agents_to_direction_col(*self)?)
-            } else if name == AgentStateField::Shape.name() {
-                Arc::new(json_vals_to_utf8(vals, true)?)
-            } else if name == AgentStateField::Height.name() {
-                Arc::new(json_vals_to_primitive::<datatypes::Float64Type>(
-                    vals, true,
-                )?)
-            } else if name == AgentStateField::Scale.name() {
-                Arc::new(agents_to_scale_col(*self)?)
-            } else if name == AgentStateField::Color.name() {
-                Arc::new(json_vals_to_utf8(vals, true)?)
-            } else if name == AgentStateField::Rgb.name() {
-                Arc::new(agents_to_rgb_col(*self)?)
-            } else if name == AgentStateField::Hidden.name() {
-                Arc::new(json_vals_to_bool(vals)?)
-            } else if name == PREVIOUS_INDEX_FIELD_KEY {
-                previous_index_to_empty_col(self.len(), field.data_type().clone())?
-            } else if matches!(
-                schema
-                    .field_spec_map
-                    .get_field_spec(&RootFieldKey::new(name))?
-                    .inner
-                    .field_type
-                    .variant,
-                FieldTypeVariant::AnyType
-            ) {
-                // Any-type (JSON string) column
-                json_vals_to_any_type_col(vals, field.data_type())?
-            } else {
-                json_vals_to_col(vals, field, field.is_nullable())?
-            };
-            cols.push(col);
-        }
-        RecordBatch::try_new(schema.arrow.clone(), cols).map_err(Error::from)
-    }
-}
 
 /// Conversion into `Agent`, which can be converted to JSON
 pub trait IntoAgents {
@@ -691,8 +462,10 @@ impl IntoAgents for RecordBatch {
 
 #[cfg(test)]
 pub mod tests {
+    use stateful::agent::arrow::array::IntoRecordBatch;
+
     use super::*;
-    use crate::datastore::test_utils::gen_schema_and_test_agents;
+    use crate::datastore::{test_utils::gen_schema_and_test_agents, Result};
 
     #[test]
     fn agent_state_into_record_batch() -> Result<()> {
@@ -703,10 +476,10 @@ pub mod tests {
             let initial_seed = round * num_agents;
             let (schema, mut agents) = gen_schema_and_test_agents(num_agents, initial_seed as u64)?;
 
-            let agent_batch = agents.as_slice().into_agent_batch(&schema)?;
+            let agent_batch = agents.as_slice().to_agent_batch(&schema)?;
             let message_batch = agents
                 .as_slice()
-                .into_message_batch(&MessageSchema::default().arrow)?;
+                .to_message_batch(&MessageSchema::default().arrow)?;
 
             let mut returned_agents =
                 (&agent_batch, &message_batch).into_agent_states(Some(&schema))?;
