--- packages/engine/src/datastore/arrow/message.rs
@@ -1,22 +1,0 @@
-use std::sync::Arc;
-
-use arrow::{array::Array, datatypes::Schema, record_batch::RecordBatch};
-use stateful::message::arrow::array::MessageArray;
-
-use crate::datastore::error::{Error, Result};
-
-pub fn batch_from_json(
-    schema: &Arc<Schema>,
-    ids: Vec<&str>,
-    messages: Option<Vec<serde_json::Value>>,
-) -> Result<RecordBatch> {
-    let agent_count = ids.len();
-    let ids = Arc::new(super::batch_conversion::get_agent_id_array(ids)?);
-
-    let messages: Arc<dyn Array> = messages.map_or_else(
-        || MessageArray::new(agent_count).map(Arc::new),
-        |values| MessageArray::from_json(values).map(Arc::new),
-    )?;
-
-    RecordBatch::try_new(schema.clone(), vec![ids, messages]).map_err(Error::from)
-}
