+++ packages/engine/src/datastore/arrow/mod.rs
@@ -1,5 +1,4 @@
 pub mod batch_conversion;
-pub mod message;
 
 #[cfg(test)]
 mod test;
